Created Login function
Users can now login to the page if their credentials match those stored in the Users table. Navigation bar icons change to reflect a logged in account, along with the homepage text. Styling has also been created for headings and images, which have been used to design the login page and new homepage.