Created review storage
Reviews can now be posted by users on a new postreview.html page. The information from the user, including optional uploaded images, is stored in the Reviews table. Images are converted to the BLOB format for SQL storage. Currently, images are displayed as BLOBs on the homepage, but will be converted in a later commit.

- Icons for the app were also added to the directory.