Added post deleting function
myreviews.html now includes options for editing (currently unused) and deleting posts. A confirmation box appears before the user confirms their choice of deleting a review, which in turn removes that review from the Reviews table and reloads the page.

Note that the CSP was changed to allow inline scripts, required for the confirmation box.