Implemented responsiveness
The login and register pages were updated to be centred with better sizing that scales to all screens. Divs with overflow styling allow for tables and long headings to be scrolled on smaller screens. The nav bar now scales correctly for any screen size with the position: sticky; styling option.